+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Simplified the `Checkbox` and `Radio` components. Removed all the extra markup and will now simply render an input field
 - Removed the `options` prop from the `Select` component. Can now be used the same ways as the native select.
 - Removed the `results-path` prop from the `Search` component. Use the `path` prop of the `Results` and `NoResults` components.
+- Query params that are not managed by the Store will no longer be deleted when updating the url
 
 ## 1.1.0
 - The fields used for the label and value of Checkbox/Radio/Select components